Move away from StoredState for Grafana.*Consumer and charm
We can't rely on a peer relation in the aggregator or a provider
charms, so we're left with StoredState there, but get rid of it
from Grafana and the associated Consumer libraries.

Update the tests to match. Since JSON doesn't have a literal
int type, make sure that the relation keys in the tests also are
not bare ints, and that we convert from Stored*